refactor(data-table): delete v1 — Stage 4 complete
Removes apps/web/src/components/data-table/ entirely (3510 LOC across
9 files). Swaps remaining skeleton imports in transactions/services/
contacts entry routes to data-table-v2. Adds exportValue/exportIgnore
to v2 ColumnMeta declaration so column builders that carry export
metadata keep type-checking.

Stage 4 complete. Stages 1–4 done end-to-end:
- Stage 1: extracted features/data-import, editable-trigger, page-filters
- Stage 2: 15-file v2 core (~1500 LOC, slot-based)
- Stage 3: migrated 15 callsites with FormSheets replacing DraftRow
- Stage 4: deleted v1, boundaries + typecheck green
