--- apps/web/src/components/data-table/data-table-empty-state.tsx
@@ -1,22 +1,0 @@
-import { useIsomorphicLayoutEffect } from "foxact/use-isomorphic-layout-effect";
-import type React from "react";
-import { useDataTable, useDataTableStore } from "./data-table-root";
-
-interface DataTableEmptyStateProps {
-   children: React.ReactNode;
-}
-
-export function DataTableEmptyState({ children }: DataTableEmptyStateProps) {
-   const { store, table, isDraftRowActive } = useDataTable();
-   const importState = useDataTableStore((s) => s.importState);
-
-   useIsomorphicLayoutEffect(() => {
-      store.setState((s) => ({ ...s, hasEmptyState: true }));
-      return () => store.setState((s) => ({ ...s, hasEmptyState: false }));
-   }, [store]);
-
-   if (table.getCoreRowModel().rows.length > 0) return null;
-   if (isDraftRowActive) return null;
-   if (importState !== null) return null;
-   return <>{children}</>;
-}
