--- apps/web/src/components/data-table/data-table-pagination.tsx
@@ -1,149 +1,0 @@
-import { useMemo } from "react";
-import {
-   Pagination,
-   PaginationContent,
-   PaginationItem,
-   PaginationLink,
-   PaginationNext,
-   PaginationPrevious,
-} from "@packages/ui/components/pagination";
-import {
-   Select,
-   SelectContent,
-   SelectItem,
-   SelectTrigger,
-   SelectValue,
-} from "@packages/ui/components/select";
-import { cn } from "@packages/ui/lib/utils";
-
-export interface DataTablePaginationProps {
-   currentPage: number;
-   totalPages: number;
-   totalCount: number;
-   pageSize: number;
-   onPageChange: (page: number) => void;
-   onPageSizeChange?: (size: number) => void;
-   pageSizeOptions?: number[];
-}
-
-const DEFAULT_PAGE_SIZE_OPTIONS = [10, 20, 30, 50, 100];
-
-function getPageNumbers(currentPage: number, totalPages: number): number[] {
-   if (totalPages <= 5)
-      return Array.from({ length: totalPages }, (_, i) => i + 1);
-   if (currentPage <= 3) return [1, 2, 3, 4, 5];
-   if (currentPage >= totalPages - 2)
-      return Array.from({ length: 5 }, (_, i) => totalPages - 4 + i);
-   return Array.from({ length: 5 }, (_, i) => currentPage - 2 + i);
-}
-
-export function DataTablePagination({
-   currentPage,
-   totalPages,
-   totalCount,
-   pageSize,
-   onPageChange,
-   onPageSizeChange,
-   pageSizeOptions = DEFAULT_PAGE_SIZE_OPTIONS,
-}: DataTablePaginationProps) {
-   if (totalCount === 0) return null;
-   const isFirstPage = currentPage === 1;
-   const isLastPage = currentPage === totalPages || totalPages === 0;
-   const hasSinglePage = totalPages <= 1;
-   const pageNumbers = useMemo(
-      () => getPageNumbers(currentPage, totalPages),
-      [currentPage, totalPages],
-   );
-
-   return (
-      <div className="sticky bottom-0 z-20 flex flex-col items-center justify-between gap-4 bg-background py-2 sm:flex-row">
-         <div className="flex items-center gap-4">
-            <div className="text-sm text-muted-foreground hidden md:block">
-               Exibindo {totalCount} resultados
-            </div>
-            <div className="flex items-center justify-center text-sm font-medium">
-               {`Página ${currentPage} de ${totalPages}`}
-            </div>
-         </div>
-         <div className="flex items-center gap-4">
-            {onPageSizeChange && (
-               <div className="flex items-center gap-2">
-                  <span className="text-sm text-muted-foreground hidden sm:inline">
-                     Linhas por página
-                  </span>
-                  <Select
-                     onValueChange={(value) => onPageSizeChange(Number(value))}
-                     value={String(pageSize)}
-                  >
-                     <SelectTrigger className="h-8 w-auto">
-                        <SelectValue placeholder={String(pageSize)} />
-                     </SelectTrigger>
-                     <SelectContent side="top">
-                        {pageSizeOptions.map((size) => (
-                           <SelectItem key={size} value={String(size)}>
-                              {size}
-                           </SelectItem>
-                        ))}
-                     </SelectContent>
-                  </Select>
-               </div>
-            )}
-            <Pagination className="w-auto">
-               <PaginationContent>
-                  <PaginationItem>
-                     <PaginationPrevious
-                        aria-disabled={isFirstPage || hasSinglePage}
-                        className={cn(
-                           (isFirstPage || hasSinglePage) &&
-                              "pointer-events-none opacity-50",
-                        )}
-                        href="#"
-                        tabIndex={isFirstPage || hasSinglePage ? -1 : undefined}
-                        onClick={(e) => {
-                           e.preventDefault();
-                           if (!isFirstPage && !hasSinglePage)
-                              onPageChange(currentPage - 1);
-                        }}
-                     />
-                  </PaginationItem>
-                  {pageNumbers.map((pageNum) => (
-                     <PaginationItem key={pageNum}>
-                        <PaginationLink
-                           aria-disabled={hasSinglePage}
-                           className={cn(
-                              hasSinglePage && "pointer-events-none opacity-50",
-                           )}
-                           href="#"
-                           isActive={pageNum === currentPage}
-                           tabIndex={hasSinglePage ? -1 : undefined}
-                           onClick={(e) => {
-                              e.preventDefault();
-                              if (!hasSinglePage) onPageChange(pageNum);
-                           }}
-                        >
-                           {pageNum}
-                        </PaginationLink>
-                     </PaginationItem>
-                  ))}
-                  <PaginationItem>
-                     <PaginationNext
-                        aria-disabled={isLastPage || hasSinglePage}
-                        className={cn(
-                           (isLastPage || hasSinglePage) &&
-                              "pointer-events-none opacity-50",
-                        )}
-                        href="#"
-                        tabIndex={isLastPage || hasSinglePage ? -1 : undefined}
-                        onClick={(e) => {
-                           e.preventDefault();
-                           if (!isLastPage && !hasSinglePage)
-                              onPageChange(currentPage + 1);
-                        }}
-                     />
-                  </PaginationItem>
-               </PaginationContent>
-            </Pagination>
-         </div>
-      </div>
-   );
-}
